Fixed DataSource's deleteById method (for some reason, it was written to accept "id" as an Integer object, and then it initialized "whereArgs" to "new String[id]", and all this did was make a string array of size "id").
Implemented Delete buttons in Update Goals screen. Will implement Delete buttons in other Update screens tomorrow.

Added an Allowance column to the User table, so that purchases can be directly subtracted from that allowance, and so that the Allowance calculations only have to be done once per day.
THIS IS NOT DONE YET.